Move imported stuff around.
