+++ tests/test_downloadermiddleware_redirect_base.py
@@ -1,7 +1,5 @@
 from __future__ import annotations
 
-from itertools import product
-
 import pytest
 
 from scrapy.downloadermiddlewares.httpproxy import HttpProxyMiddleware
@@ -9,43 +7,6 @@ from scrapy.exceptions import IgnoreRequest
 from scrapy.http import Request, Response
 from scrapy.utils.misc import set_environ
 from scrapy.utils.test import get_crawler
-
-SCHEME_PARAMS = ("url", "location", "target")
-HTTP_SCHEMES = ("http", "https")
-NON_HTTP_SCHEMES = ("data", "file", "ftp", "s3", "foo")
-REDIRECT_SCHEME_CASES = (
-    # http/https → http/https redirects
-    *(
-        (
-            f"{input_scheme}://example.com/a",
-            f"{output_scheme}://example.com/b",
-            f"{output_scheme}://example.com/b",
-        )
-        for input_scheme, output_scheme in product(HTTP_SCHEMES, repeat=2)
-    ),
-    # http/https → data/file/ftp/s3/foo does not redirect
-    *(
-        (
-            f"{input_scheme}://example.com/a",
-            f"{output_scheme}://example.com/b",
-            None,
-        )
-        for input_scheme in HTTP_SCHEMES
-        for output_scheme in NON_HTTP_SCHEMES
-    ),
-    # http/https → relative redirects
-    *(
-        (
-            f"{scheme}://example.com/a",
-            location,
-            f"{scheme}://example.com/b",
-        )
-        for scheme in HTTP_SCHEMES
-        for location in ("//example.com/b", "/b")
-    ),
-    # Note: We do not test data/file/ftp/s3 schemes for the initial URL
-    # because their download handlers cannot return a status code of 3xx.
-)
 
 
 class Base:
