+++ tests/test_scheduler.py
@@ -2,7 +2,6 @@ from __future__ import annotations
 
 import warnings
 from abc import ABC, abstractmethod
-from collections import deque
 from contextlib import AbstractAsyncContextManager, asynccontextmanager
 from typing import TYPE_CHECKING, Any, NamedTuple, cast
 from unittest.mock import Mock
@@ -10,7 +9,7 @@ from unittest.mock import Mock
 import pytest
 
 from scrapy.core.downloader import Downloader
-from scrapy.core.scheduler import BaseScheduler, Scheduler
+from scrapy.core.scheduler import Scheduler
 from scrapy.crawler import Crawler
 from scrapy.exceptions import ScrapyDeprecationWarning
 from scrapy.http import Request
@@ -25,38 +24,6 @@ from tests.utils.decorators import coroutine_test, inline_callbacks_test
 if TYPE_CHECKING:
     from collections.abc import AsyncGenerator
     from pathlib import Path
-
-
-class MemoryScheduler(BaseScheduler):
-    paused = False
-
-    def __init__(self, *args: Any, **kwargs: Any):
-        super().__init__(*args, **kwargs)
-        self.queue: deque[Request] = deque(
-            Request(value) if isinstance(value, str) else value
-            for value in getattr(self, "queue", [])
-        )
-
-    def enqueue_request(self, request: Request) -> bool:
-        self.queue.append(request)
-        return True
-
-    def has_pending_requests(self) -> bool:
-        return self.paused or bool(self.queue)
-
-    def next_request(self) -> Request | None:
-        if self.paused:
-            return None
-        try:
-            return self.queue.pop()
-        except IndexError:
-            return None
-
-    def pause(self) -> None:
-        self.paused = True
-
-    def unpause(self) -> None:
-        self.paused = False
 
 
 class MockSlot(NamedTuple):
